Reorganize auto-start section of User Guide
The how-to section on auto-running bash files was deleted from the amdgpu-pac section, lightly edited, and moved to a separate section at the end of the Guide.  Also fixed a spelling typo, 'installation', at paragraph #52. And at para # 164, removed the driver warning message in the amdgpu-ls output example because it no longer displays in the most recent Master update.