update to User Guide
Deleted sentence at line 261 because the issue is resolved in current Master.
Changed crontab commands to a more proper execution (if reading of first bash file fails, second file will still be read).
Added to the final autostart section instructions for setting up a systemd startup service.  The written order of the commands is how I did it, but there is probably a more logical or more Linuxy way to go about it - please edit as needed.