Mention that the package is in the main repos since Bullseye
Hi,

I really like the project, with Corectrl being such a nightmare to compile on Debian, this is an excellent, few-dependencies alternative. However, I think it'd be worthwhile to mention that it can be installed from the main repos if one uses Bullseye (which is what most users will be soon enough). For this reason I've added a tiny edit in the README.

Have a nice day!